+++ test/Ether.Network.Tests/Server/TestClient.cs
@@ -1,0 +1,22 @@
+﻿using Ether.Network.Packets;
+
+namespace Ether.Network.Tests.Server
+{
+    internal class TestClient : NetConnection
+    {
+        public void SendHello()
+        {
+            using (var packet = new NetPacket())
+            {
+                packet.Write(0); // Hello header
+                packet.Write(NetServerTest.WelcomeText);
+
+                this.Send(packet);
+            }
+        }
+
+        public override void HandleMessage(NetPacketBase packet)
+        {
+        }
+    }
+}
